# Optimize load of ERC20 addresses