- Closes #2301

Use database iterators to load ERC20 addresses, so only chunks are loaded in RAM, instead of the whole queryset. Using production database I analyzed good values for `chunk_size` value:

```
        Chunk size optimization
        -----------------------
        Testing with 3M Safes
        2k - 90 seconds
        100k - 73 seconds
        500k - 60 seconds
        1M - 60 seconds
        3M - 60 seconds
        
        Testing with 15M Safes
        50k - 854 seconds
        500k - 460 seconds
        750k - 415 seconds
        1M - 430 seconds
        2M - 398 seconds
        3M - 407 seconds
        
        500k sounds like a good compromise memory/speed wise
```